Feat: 모임 참가하기 (#95)
* Feat: 모임 참가자 대기 엔티티 및 레포지토리 생성

- MeetingParticipateWaitingEntity 생성
- MeetingParticipantSubEntity participantCount 필드 타입 변경

* Feat: 모임참가 기능 구현

* Chore: 불필요 코드 삭제

* Fix: 모임참가하기 비즈니스 로직 수정

- 모임 참가하기 요청데이터 유효성 검증 추가

* Feat: 모임참가하기 RestDocs 테스트코드 작성

* Fix: MeetingParticipantSubEntity participantCount 필드 타입변환 사이드이펙트

* Feat: 모임참가하기 RestDocs 통합테스트

- meeting, meetingParticipant Fixture 생성
- 모임참가하기 API Http 응답코드 변경 (200 > 201)

* Feat: 모임 참여하기 서비스 통합테스트 코드 작성

* Chore: 미팅참가 엔티티 Fixture 메소드명 변경

* Fix: RestDocs 문서화 파일 포멧 수정

- 에러코드 테이블 작성

* Feat: 테스트 이슈 수정

* Fix: 로그백 개발,운영 파일 경로 수정

* Feat: application-dev, prod 로깅파일 경로 수정

* Fix: Archunit 네이밍 테스트 수정

* Refacor: 모임참가하기 유효성 검증 도메인으로 이동

+++ src/main/java/com/boardgo/domain/meeting/repository/projection/MeetingSearchProjection.java
@@ -13,7 +13,7 @@ public record MeetingSearchProjection(
         Integer limitParticipant,
         String nickName,
         String genres,
-        Long participantCount) {
+        Integer participantCount) {
     @QueryProjection
     public MeetingSearchProjection {}
 }
